Move pretty instance to another module
