+++ src/IR/IR.hs
@@ -23,7 +23,6 @@ module IR.IR
   , Annotation
   , Annotations
   , variantFields
-  , wellFormed
   , foldLambda
   , unfoldLambda
   , extract
@@ -43,8 +42,6 @@ import           Common.Identifiers             ( CSym(..)
                                                 , TVarId(..)
                                                 , VarId(..)
                                                 )
-import           Common.Pretty
-import           Control.Monad                  ( void )
 import           Data.Data                      ( Data
                                                 , Typeable
                                                 )
@@ -58,7 +55,6 @@ import           IR.Types.Type                  ( Annotation
                                                 , Annotations
                                                 , pattern Arrow
                                                 , Type
-                                                , unfoldArrow
                                                 )
 
 {- | Top-level compilation unit.
@@ -371,314 +367,3 @@ instance HasFreeVars (Expr t) VarId where
     freeAltVars :: (Alt t, Expr t) -> S.Set VarId
     freeAltVars (a, e) = freeVars e \\ S.fromList (mapMaybe _binderId $ altBinders a)
   freeVars Exception{} = S.empty
-
-{- | Predicate of whether an expression "looks about right".
-
-Description left deliberately vague so that we have the flexibility to
-strengthen this predicate. For now, we just check that all primitives are
-applied to the right number of arguments.
-
-This predicate also provides a template to recursively traverse through all
-sub-expressions of an expression.
--}
-wellFormed :: Expr t -> Bool
-wellFormed (Var  _ _          ) = True
-wellFormed (Data _ _          ) = True
-wellFormed (Lit  _ _          ) = True
-wellFormed (Let    defs b    _) = all (wellFormed . snd) defs && wellFormed b
-wellFormed (Lambda _    b    _) = wellFormed b
-wellFormed (App    f    a    _) = wellFormed f && wellFormed a
-wellFormed (Match  s    alts _) = wellFormed s && all (wellFormed . snd) alts
-wellFormed (Prim   p    es   _) = wfPrim p es && all wellFormed es
- where
-  wfPrim New                  [_]       = True
-  wfPrim Dup                  [_]       = True
-  wfPrim Drop                 [_, _]    = True
-  wfPrim Deref                [_]       = True
-  wfPrim Assign               [_, _]    = True
-  wfPrim After                [_, _, _] = True
-  wfPrim Par                  (_ : _)   = True
-  wfPrim Wait                 (_ : _)   = True
-  wfPrim Break                []        = True
-  wfPrim Now                  [_]       = True
-  wfPrim (CCall   _         ) _         = True
-  wfPrim (FfiCall _         ) [_]       = True
-  wfPrim (PrimOp  PrimNeg   ) [_]       = True
-  wfPrim (PrimOp  PrimNot   ) [_]       = True
-  wfPrim (PrimOp  PrimBitNot) [_]       = True
-  wfPrim (PrimOp  PrimAdd   ) [_, _]    = True
-  wfPrim (PrimOp  PrimSub   ) [_, _]    = True
-  wfPrim (PrimOp  PrimMul   ) [_, _]    = True
-  wfPrim (PrimOp  PrimDiv   ) [_, _]    = True
-  wfPrim (PrimOp  PrimMod   ) [_, _]    = True
-  wfPrim (PrimOp  PrimBitAnd) [_, _]    = True
-  wfPrim (PrimOp  PrimBitOr ) [_, _]    = True
-  wfPrim (PrimOp  PrimEq    ) [_, _]    = True
-  wfPrim (PrimOp  PrimGt    ) [_, _]    = True
-  wfPrim (PrimOp  PrimGe    ) [_, _]    = True
-  wfPrim (PrimOp  PrimLt    ) [_, _]    = True
-  wfPrim (PrimOp  PrimLe    ) [_, _]    = True
-  wfPrim _                    _         = False
-wellFormed (Exception _ _) = True
-
-{- | Pretty Typeclass: pretty print the IR
-
-Adds 
-* indentation and line breaks
-* some parens (not minimal parens, but fewer than around every node)
-Omits 
-* let _ =
-* type annotations
-Reverts
-* curried funcs of one arg back to multiple arg funcs
--}
-instance Pretty (Program Type) where
-  pretty Program { programDefs = ds, typeDefs = tys, externDecls = xds } =
-    vsep $ punctuate line tops
-   where
-    tops =
-      map prettyTypDef tys ++ map prettyExternDecl xds ++ map prettyFuncDef ds
-
-    -- Generates readable Doc representation of an IR Top Level Function
-    prettyFuncDef :: (VarId, Expr Type) -> Doc ann
-    prettyFuncDef (v, l@(Lambda _ _ ty)) =
-      pretty v <+> typSig <+> pretty "=" <+> line <> indent
-        2
-        (pretty (void body))
-     where
-      typSig = hsep args <+> rarrow <+> pretty retTy
-      args   = zipWith
-        (\arg t -> parens $ pretty arg <+> pretty ":" <+> pretty t)
-        argIds
-        argTys
-      (argIds, body ) = unfoldLambda l
-      (argTys, retTy) = unfoldArrow ty -- FIXME
-    prettyFuncDef (v, e) = pretty v <+> pretty "=" <+> pretty (void e)
-
-    prettyExternDecl :: (Pretty t) => (VarId, t) -> Doc ann
-    prettyExternDecl (v, t) =
-      pretty "extern" <+> pretty v <+> colon <+> pretty t
-
-    -- Generates readable Doc representation of an IR Type Definition
-    prettyTypDef :: (TConId, TypeDef) -> Doc ann
-    prettyTypDef (tcon, TypeDef { variants = vars }) =
-      pretty "type"
-        <+> pretty tcon
-        <+> line
-        <>  indent indentNo (vsep $ map prettyDCon vars)
-        <>  line
-    prettyDCon :: (DConId, TypeVariant) -> Doc ann
-    prettyDCon (dcon, VariantNamed argz) =
-      pretty dcon <+> hsep (pretty . snd <$> argz)
-    prettyDCon (dcon, VariantUnnamed argz) =
-      pretty dcon <+> hsep (pretty <$> argz)
-
-
--- | Pretty prints IR Expr nodes without type annotations
-instance Pretty (Expr ()) where
-  pretty = undefined
-    {-
-  pretty a@App{} = pretty nm <+> hsep (parenz . fst <$> args)
-   where
-    (nm, args) = unfoldApp a
-    -- insert (usually) necessary parens
-    parenz :: Expr () -> Doc ann
-    parenz v@(Var _ _) = pretty v  -- variables
-    parenz l@(Lit _ _) = pretty l  -- literals
-    parenz e           = parens (pretty e)
-    -- TODO: minimum parens algo
-  pretty (Prim Wait es _           ) = pretty "wait" <+> vsep (map pretty es)
-  pretty (Var v _                  ) = pretty v
-  pretty (Lambda a              b _) = pretty "fun" <+> pretty a <+> pretty b
-  -- pretty (Let    [(Nothing, e)] b _) = pretty e <> line <> pretty b
-  pretty (Let    as             b _) = letexpr
-   where
-    letexpr = pretty "let" <+> vsep (map def as) <> line <> pretty b
-    def (_binderId -> Just v, e) = pretty v <+> equals <+> align (pretty e)
-    def (_, e) = pretty '_' <+> equals <+> align (pretty e)
-  pretty (Prim After [d, l, r] _) = ae
-   where
-    ae =
-      pretty "after" <+> pretty d <> comma <+> pretty l <+> larrow <+> pretty r
-  pretty (Prim  Assign [l, r] _) = parens $ pretty l <+> larrow <+> pretty r
-  pretty (Match s      as     _) = pretty "match" <+> pretty s <> line <> arms
-   where
-    arms = vsep (map (indent indentNo . arm) as)
-    arm :: (Alt t, Expr ()) -> Doc ann
-    arm (a, e) = pretty a <+> equals <+> align (pretty e)
-  pretty (Prim Loop [b] _) =
-    pretty "loop" <> line <> indent indentNo (pretty b)
-  pretty (Prim (PrimOp po) [l, r] _) = pretty l <+> pretty po <+> pretty r
-  pretty (Data d _                 ) = pretty d
-  pretty (Lit  l _                 ) = pretty l
-  pretty (Prim New [r] _           ) = pretty "new" <+> pretty r
-  pretty (Prim Dup [r] _           ) = pretty "__dup" <+> parens (pretty r)
-  pretty (Prim Drop [e, r] _) =
-    pretty "__drop"
-      <+> parens (line <> indent 2 (pretty e) <> line)
-      <+> pretty r
-  pretty (Prim Deref [r] _) = pretty "deref" <+> parens (pretty r)
-  pretty (Prim Par   es  _) = pretty "par" <+> block dbar (map pretty es)
-  pretty (Prim Break []  _) = pretty "break"
-  pretty (Prim (CCall s) es _) =
-    pretty "$" <> pretty s <> parens (hsep $ punctuate comma $ map pretty es)
-  pretty (Prim (FfiCall s) es _) = pretty s <+> hsep (map (parens . pretty) es)
-  pretty (Prim (CQuote  s) [] _) = pretty "$$" <> pretty s <> pretty "$$"
-
-  -- pretty (Prim Return [e] _        ) = pretty "return" <+> braces (pretty e)
-  pretty (Prim p _ _) = error "Primitive expression not well-formed: " $ show p
-  pretty (Exception et _       ) = pretty "__exception" <+> prettyet et
-    where prettyet (ExceptDefault l) = pretty l
-    -}
-
-instance Pretty (Alt Type) where
-  pretty = undefined
-  -- pretty (AltData a b       _ ) = pretty a <+> hsep (map pretty b)
-  -- pretty (AltLit    a       _ ) = pretty a
-  -- pretty (AltBinder (_binderId -> Just v)) = pretty v
-  -- pretty (AltBinder (_binderId -> Nothing)) = pretty '_'
-
-instance Pretty Literal where
-  pretty = undefined
-  -- pretty (LitIntegral i) = pretty $ show i
-  -- pretty LitEvent        = pretty "()"
-
-instance Pretty PrimOp where
-  pretty PrimNeg    = pretty "-"
-  pretty PrimNot    = pretty "!"
-  pretty PrimBitNot = pretty "~"
-  pretty PrimAdd    = pretty "+"
-  pretty PrimSub    = pretty "-"
-  pretty PrimMul    = pretty "*"
-  pretty PrimDiv    = pretty "/"
-  pretty PrimMod    = pretty "%"
-  pretty PrimBitAnd = pretty "&"
-  pretty PrimBitOr  = pretty "|"
-  pretty PrimEq     = pretty "=="
-  pretty PrimNeq    = pretty "!="
-  pretty PrimGt     = pretty ">"
-  pretty PrimGe     = pretty ">="
-  pretty PrimLt     = pretty "<"
-  pretty PrimLe     = pretty "<="
-
--- | Dumpy Typeclass: generate comprehensive Doc representation of the IR
-instance Dumpy (Program Type) where
-  dumpy = undefined
-  -- dumpy Program { programDefs = ds, typeDefs = tys, externDecls = xds } =
-  --   vsep $ punctuate line tops
-  --  where
-  --   tops =
-  --     map dumpyTypDef tys ++ map dumpyExternDecl xds ++ map dumpyFuncDef ds
-  --
-  --   -- Generates readable Doc representation of an IR Top Level Function
-  --   dumpyFuncDef :: (VarId, Expr Type) -> Doc ann
-  --   dumpyFuncDef (v, l@(Lambda _ _ ty)) =
-  --     pretty v <+> typSig <+> pretty "=" <+> line <> indent 2 (dumpy body)
-  --    where
-  --     typSig = hsep args <+> rarrow <+> dumpy retTy
-  --     args   = zipWith
-  --       (\arg t -> parens $ pretty arg <+> pretty ":" <+> dumpy t)
-  --       argIds
-  --       argTys
-  --     (argIds, body ) = unfoldLambda l
-  --     (argTys, retTy) = unfoldArrow ty
-  --   dumpyFuncDef (v, e) = pretty v <+> pretty "=" <+> dumpy e
-  --
-  --   -- Generates readable Doc representation of an IR Type Definition
-  --   dumpyExternDecl :: Dumpy t => (VarId, t) -> Doc ann
-  --   dumpyExternDecl (v, t) = pretty "extern" <+> pretty v <+> colon <+> dumpy t
-  --
-  --   -- Generates readable Doc representation of an IR Type Definition
-  --   dumpyTypDef :: (TConId, TypeDef) -> Doc ann
-  --   dumpyTypDef (tcon, TypeDef { variants = vars }) =
-  --     pretty "type"
-  --       <+> pretty tcon
-  --       <+> line
-  --       <>  indent indentNo (vsep $ map dumpyDCon vars)
-  --       <>  line
-  --   dumpyDCon :: (DConId, TypeVariant) -> Doc ann
-  --   dumpyDCon (dcon, VariantNamed argz) =
-  --     pretty dcon <+> hsep (dumpy . snd <$> argz)
-  --   dumpyDCon (dcon, VariantUnnamed argz) =
-  --     pretty dcon <+> hsep (dumpy <$> argz)
-  -- -- TODO: how to represent entry point?
-
-instance Pretty (Binder Type) where
-  pretty Binder {_binderId = Just v} = pretty v
-  pretty Binder {_binderId = Nothing } = pretty '_'
-
-instance Dumpy (Binder Type) where
-  dumpy = pretty
-
-instance Dumpy (Expr Type) where
-  dumpy (Var  v t  ) = typeAnn t $ pretty v
-  dumpy (Data d t  ) = typeAnn t $ pretty d
-  dumpy (Lit  l t  ) = typeAnn t $ dumpy l
-  dumpy (App l  r t) = typeAnn t $ dumpy l <+> dumpy r
-  dumpy (Let as b t) = typeAnn t $ parens letexpr
-   where
-    letexpr = pretty "let" <+> block dbar (map def as) <> semi <+> dumpy b
-    def (_binderId -> Just v , e) = pretty v <+> equals <+> braces (dumpy e)
-    def (_, e) = pretty '_' <+> equals <+> braces (dumpy e)
-  dumpy (Lambda a b t) =
-    typeAnn t $ pretty "fun" <+> pretty a <+> braces (dumpy b)
-  dumpy (Match s as t) = typeAnn t $ pretty "match" <+> dumpy s <+> arms
-   where
-    -- Where to add binder?
-    arms = block bar (map arm as)
-    arm (a, e) = dumpy a <+> pretty "=" <+> braces (dumpy e)
-  dumpy (Prim New [r] t) = typeAnn t $ pretty "new" <+> dumpy r
-  dumpy (Prim Dup [r] t) = typeAnn t $ pretty "__dup" <+> dumpy r
-  dumpy (Prim Drop [e, r] t) =
-    typeAnn t $ pretty "__drop" <+> parens (dumpy e) <+> parens (dumpy r)
-  dumpy (Prim Deref [r] t) = typeAnn t $ pretty "deref" <+> dumpy r
-  dumpy (Prim Assign [l, r] t) =
-    typeAnn t $ parens $ dumpy l <+> larrow <+> braces (dumpy r)
-  dumpy (Prim After [d, l, r] t) = typeAnn t $ parens ae
-   where
-    ae = pretty "after" <+> dumpy d <> comma <+> dumpy l <+> larrow <+> braces
-      (dumpy r)
-  dumpy (Prim Par es t) =
-    typeAnn t $ pretty "par" <+> block dbar (map dumpy es)
-  dumpy (Prim Wait es t) =
-    typeAnn t $ pretty "wait" <+> block dbar (map dumpy es)
-  dumpy (Prim Loop      [b] t) = typeAnn t $ pretty "loop" <+> braces (dumpy b)
-  dumpy (Prim Break     []  t) = typeAnn t $ pretty "break"
-  dumpy (Prim (CCall s) es  t) = typeAnn t $ pretty "$" <> pretty s <> parens
-    (hsep $ punctuate comma $ map dumpy es)
-  dumpy (Prim (CQuote s) [] t) =
-    typeAnn t $ pretty "$$" <> pretty s <> pretty "$$"
-  dumpy (Prim (FfiCall s) es _) = pretty s <+> hsep (map (parens . dumpy) es)
-  dumpy (Prim (PrimOp po) [l, r] t) =
-    typeAnn t $ dumpy l <+> dumpy po <+> dumpy r
-  dumpy (Prim p _ _) = error "Primitive expression not well-formed: " $ show p
-  dumpy (Exception et t) = typeAnn t $ pretty "exception" <+> dumpyet et
-    where dumpyet (ExceptDefault l) = pretty l
-
-instance Dumpy (Alt Type) where 
-  dumpy (AltData a b      _) = parens $ pretty a <+> hsep (map pretty b)
-  dumpy (AltLit    a      _) = pretty a
-  dumpy (AltBinder (_binderId -> Just v))        = pretty v
-  dumpy (AltBinder _)        = pretty '_'
-
-instance Dumpy Literal where
-  dumpy (LitIntegral i) = pretty $ show i
-  dumpy LitEvent        = pretty "()"
-
-instance Dumpy PrimOp where
-  dumpy PrimNeg    = pretty "-"
-  dumpy PrimNot    = pretty "!"
-  dumpy PrimBitNot = pretty "~"
-  dumpy PrimAdd    = pretty "+"
-  dumpy PrimSub    = pretty "-"
-  dumpy PrimMul    = pretty "*"
-  dumpy PrimDiv    = pretty "/"
-  dumpy PrimMod    = pretty "%"
-  dumpy PrimBitAnd = pretty "&"
-  dumpy PrimBitOr  = pretty "|"
-  dumpy PrimEq     = pretty "=="
-  dumpy PrimNeq    = pretty "!="
-  dumpy PrimGt     = pretty ">"
-  dumpy PrimGe     = pretty ">="
-  dumpy PrimLt     = pretty "<"
-  dumpy PrimLe     = pretty "<="
